[DLv2] Move serialization functions to graph
