+++ torchdata/dataloader2/dataloader2.py
@@ -5,21 +5,17 @@
 # LICENSE file in the root directory of this source tree.
 
 
-import pickle
 from dataclasses import dataclass
 from typing import Any, Dict, Generic, Iterable, Iterator, Optional, TypeVar, Union
 
-from torch.utils.data.datapipes.datapipe import (
-    _DataPipeSerializationWrapper,
-    _IterDataPipeSerializationWrapper,
-    _MapDataPipeSerializationWrapper,
-)
-
-from torch.utils.data.graph import DataPipe
 from torchdata.dataloader2.adapter import Adapter
-
-from torchdata.datapipes.iter import IterDataPipe
-from torchdata.datapipes.map import MapDataPipe
+from torchdata.dataloader2.graph import (
+    clone,
+    DataPipe,
+    deserialize_datapipe,
+    serialize_datapipe,
+    wrap_datapipe_for_serialization,
+)
 
 from .error import PauseIteration
 from .reading_service import CheckpointableReadingServiceInterface, ReadingServiceInterface
@@ -27,20 +23,6 @@ from .reading_service import CheckpointableReadingServiceInterface, ReadingServi
 T_co = TypeVar("T_co", covariant=True)
 SERIALIZED_DATAPIPE_KEY_NAME = "serialized_datapipe"
 READING_SERVICE_STATE_KEY_NAME = "reading_service_state"
-
-
-def serialize_datapipe(datapipe: DataPipe) -> bytes:
-    try:
-        return pickle.dumps(datapipe)
-    except pickle.PickleError as e:
-        raise NotImplementedError(f"Prototype only support pickle-able datapipes for checkpoint: {e}")
-
-
-def deserialize_datapipe(serialized_state: bytes) -> DataPipe:
-    try:
-        return pickle.loads(serialized_state)
-    except pickle.PickleError as e:
-        raise NotImplementedError(f"Prototype only support pickle-able datapipes for checkpoint: {e}")
 
 
 @dataclass
@@ -128,7 +110,7 @@ class DataLoader2(Generic[T_co]):
         datapipe_adapter_fn: Optional[Union[Iterable[Adapter], Adapter]] = None,
         reading_service: Optional[ReadingServiceInterface] = None,
     ) -> None:
-        self.datapipe = self._wrap_and_copy_dp(datapipe) if datapipe is not None else None
+        self.datapipe = clone(wrap_datapipe_for_serialization(datapipe)) if datapipe is not None else None
         self._adapted: bool = False
         self._datapipe_iter: Optional[Iterator[T_co]] = None
         self._reset_iter: bool = True  # Sets to `False` when __iter__ starts, and `True` when `StopIteration``
@@ -147,7 +129,7 @@ class DataLoader2(Generic[T_co]):
         if self.datapipe is not None and self.datapipe_adapter_fns is not None:
             for adapter_fn in self.datapipe_adapter_fns:
                 self.datapipe = adapter_fn(self.datapipe)
-        self._datapipe_before_reading_service_adapt: DataPipe = self._copy(self.datapipe)
+        self._datapipe_before_reading_service_adapt: DataPipe = clone(self.datapipe)
 
     def __iter__(self) -> DataLoader2Iterator[T_co]:
         r"""
@@ -183,29 +165,6 @@ class DataLoader2(Generic[T_co]):
 
     def __del__(self) -> None:
         self.shutdown()
-
-    @staticmethod
-    def _copy(obj):
-        r"""
-        Standardized way for DataLoader2 to copy an object when needed, such as for DataPipe/ReadingService.
-        This uses `pickle` to serialize/deserialize to create the copy.
-        """
-        return pickle.loads(pickle.dumps(obj))
-
-    @staticmethod
-    def _wrap_and_copy_dp(datapipe: DataPipe):
-        r"""
-        Wraps the ``DataPipe`` with the corresponding serialization wrapper.
-        Then, creates a copy with the class's static copy method.
-
-        """
-        wrapped_dp: DataPipe = datapipe
-        if not isinstance(datapipe, _DataPipeSerializationWrapper):
-            if isinstance(datapipe, IterDataPipe):
-                wrapped_dp = _IterDataPipeSerializationWrapper(datapipe)
-            elif isinstance(datapipe, MapDataPipe):
-                wrapped_dp = _MapDataPipeSerializationWrapper(datapipe)
-        return DataLoader2._copy(wrapped_dp)
 
     def shutdown(self) -> None:
         r"""
@@ -294,4 +253,4 @@ class DataLoader2(Generic[T_co]):
         if self.datapipe_adapter_fns is not None:
             for adapter_fn in self.datapipe_adapter_fns:
                 self.datapipe = adapter_fn(self.datapipe)
-        self._datapipe_before_reading_service_adapt = self._copy(self.datapipe)
+        self._datapipe_before_reading_service_adapt = clone(self.datapipe)
